Move Itask to loader module
