+++ mantis-control-plane/mantis-control-plane-server/dependencies.lock
@@ -70,7 +70,7 @@
             "firstLevelTransitive": [
                 "io.mantisrx:mantis-control-plane-core"
             ],
-            "locked": "2.11.2"
+            "locked": "2.12.0"
         },
         "org.apache.flink:flink-rpc-akka": {
             "firstLevelTransitive": [
@@ -215,7 +215,7 @@
             "firstLevelTransitive": [
                 "io.mantisrx:mantis-control-plane-core"
             ],
-            "locked": "2.11.2"
+            "locked": "2.12.0"
         },
         "junit:junit": {
             "locked": "4.11"
@@ -389,7 +389,7 @@
             "firstLevelTransitive": [
                 "io.mantisrx:mantis-control-plane-core"
             ],
-            "locked": "2.11.2"
+            "locked": "2.12.0"
         },
         "org.apache.flink:flink-rpc-akka": {
             "firstLevelTransitive": [
@@ -571,7 +571,7 @@
             "firstLevelTransitive": [
                 "io.mantisrx:mantis-control-plane-core"
             ],
-            "locked": "2.11.2"
+            "locked": "2.12.0"
         },
         "net.jcip:jcip-annotations": {
             "firstLevelTransitive": [
@@ -757,7 +757,7 @@
             "firstLevelTransitive": [
                 "io.mantisrx:mantis-control-plane-core"
             ],
-            "locked": "2.11.2"
+            "locked": "2.12.0"
         },
         "junit:junit": {
             "locked": "4.11"
@@ -956,7 +956,7 @@
             "firstLevelTransitive": [
                 "io.mantisrx:mantis-control-plane-core"
             ],
-            "locked": "2.11.2"
+            "locked": "2.12.0"
         },
         "junit:junit": {
             "locked": "4.11"
