+++ mantis-runtime-loader/build.gradle
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Netflix, Inc.
+ * Copyright 2019 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,18 +14,13 @@
  * limitations under the License.
  */
 
-package io.mantisrx.server.master.client;
+dependencies {
+    implementation libraries.hadoopCommon
+    api project(":mantis-control-plane:mantis-control-plane-client")
 
-import io.mantisrx.server.core.ExecuteStageRequest;
-import org.apache.flink.util.UserCodeClassLoader;
+    testImplementation libraries.junit4
+    testImplementation libraries.mockitoAll
+    testImplementation libraries.spectatorApi
+    testImplementation(testFixtures(project(":mantis-common")))
 
-/**
- * Interface to factory building ITask implementation instance. Can be override to use customized ITask impl.
- */
-public interface TaskFactory {
-    ITask getITaskInstance(ExecuteStageRequest request, ClassLoader cl);
-
-    UserCodeClassLoader getUserCodeClassLoader(
-        ExecuteStageRequest request,
-        ClassLoaderHandle classLoaderHandle);
 }
