+++ mantis-runtime-loader/src/test/java/io/mantisrx/runtime/loader/SubscriptionStateHandlerImplTest.java
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package io.mantisrx.server.master.client;
+package io.mantisrx.runtime.loader;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import io.mantisrx.common.util.DelegateClock;
-import io.mantisrx.server.master.client.SubscriptionStateHandlerImpl.SubscriptionState;
+import io.mantisrx.runtime.loader.SubscriptionStateHandlerImpl.SubscriptionState;
 import java.time.Clock;
 import java.time.Duration;
 import java.time.Instant;
