+++ mantis-server/mantis-server-worker/build.gradle
@@ -25,6 +25,7 @@ ext {
 
 dependencies {
     compile project(":mantis-runtime")
+    compile project(":mantis-runtime-loader")
     compile project(":mantis-control-plane:mantis-control-plane-client")
     compile project(":mantis-control-plane:mantis-control-plane-core")
     compile project(":mantis-server:mantis-server-worker-client")
