+++ mantis-server/mantis-server-worker/dependencies.lock
@@ -39,6 +39,7 @@
         },
         "io.mantisrx:mantis-control-plane-client": {
             "firstLevelTransitive": [
+                "io.mantisrx:mantis-runtime-loader",
                 "io.mantisrx:mantis-server-worker-client"
             ],
             "project": true
@@ -64,6 +65,9 @@
             "project": true
         },
         "io.mantisrx:mantis-runtime": {
+            "project": true
+        },
+        "io.mantisrx:mantis-runtime-loader": {
             "project": true
         },
         "io.mantisrx:mantis-rxcontrol": {
@@ -116,7 +120,7 @@
             "firstLevelTransitive": [
                 "io.mantisrx:mantis-control-plane-core"
             ],
-            "locked": "2.11.2"
+            "locked": "2.12.0"
         },
         "nz.ac.waikato.cms.moa:moa": {
             "locked": "2017.06"
@@ -262,7 +266,7 @@
             "firstLevelTransitive": [
                 "io.mantisrx:mantis-control-plane-core"
             ],
-            "locked": "2.11.2"
+            "locked": "2.12.0"
         },
         "junit:junit": {
             "locked": "4.11"
@@ -378,6 +382,7 @@
         },
         "io.mantisrx:mantis-control-plane-client": {
             "firstLevelTransitive": [
+                "io.mantisrx:mantis-runtime-loader",
                 "io.mantisrx:mantis-server-worker-client"
             ],
             "project": true
@@ -403,6 +408,9 @@
             "project": true
         },
         "io.mantisrx:mantis-runtime": {
+            "project": true
+        },
+        "io.mantisrx:mantis-runtime-loader": {
             "project": true
         },
         "io.mantisrx:mantis-rxcontrol": {
@@ -455,7 +463,7 @@
             "firstLevelTransitive": [
                 "io.mantisrx:mantis-control-plane-core"
             ],
-            "locked": "2.11.2"
+            "locked": "2.12.0"
         },
         "nz.ac.waikato.cms.moa:moa": {
             "locked": "2017.06"
@@ -593,6 +601,7 @@
         },
         "io.mantisrx:mantis-control-plane-client": {
             "firstLevelTransitive": [
+                "io.mantisrx:mantis-runtime-loader",
                 "io.mantisrx:mantis-server-worker-client"
             ],
             "project": true
@@ -618,6 +627,9 @@
             "project": true
         },
         "io.mantisrx:mantis-runtime": {
+            "project": true
+        },
+        "io.mantisrx:mantis-runtime-loader": {
             "project": true
         },
         "io.mantisrx:mantis-rxcontrol": {
@@ -680,7 +692,7 @@
             "firstLevelTransitive": [
                 "io.mantisrx:mantis-control-plane-core"
             ],
-            "locked": "2.11.2"
+            "locked": "2.12.0"
         },
         "net.jcip:jcip-annotations": {
             "firstLevelTransitive": [
@@ -711,7 +723,8 @@
         },
         "org.apache.hadoop:hadoop-common": {
             "firstLevelTransitive": [
-                "io.mantisrx:mantis-control-plane-client"
+                "io.mantisrx:mantis-control-plane-client",
+                "io.mantisrx:mantis-runtime-loader"
             ],
             "locked": "2.7.7"
         },
@@ -817,6 +830,7 @@
         },
         "io.mantisrx:mantis-control-plane-client": {
             "firstLevelTransitive": [
+                "io.mantisrx:mantis-runtime-loader",
                 "io.mantisrx:mantis-server-worker-client"
             ],
             "project": true
@@ -843,6 +857,9 @@
             "project": true
         },
         "io.mantisrx:mantis-runtime": {
+            "project": true
+        },
+        "io.mantisrx:mantis-runtime-loader": {
             "project": true
         },
         "io.mantisrx:mantis-rxcontrol": {
@@ -898,7 +915,7 @@
             "firstLevelTransitive": [
                 "io.mantisrx:mantis-control-plane-core"
             ],
-            "locked": "2.11.2"
+            "locked": "2.12.0"
         },
         "junit:junit": {
             "locked": "4.11"
@@ -1044,6 +1061,7 @@
         },
         "io.mantisrx:mantis-control-plane-client": {
             "firstLevelTransitive": [
+                "io.mantisrx:mantis-runtime-loader",
                 "io.mantisrx:mantis-server-worker-client"
             ],
             "project": true
@@ -1070,6 +1088,9 @@
             "project": true
         },
         "io.mantisrx:mantis-runtime": {
+            "project": true
+        },
+        "io.mantisrx:mantis-runtime-loader": {
             "project": true
         },
         "io.mantisrx:mantis-rxcontrol": {
@@ -1135,7 +1156,7 @@
             "firstLevelTransitive": [
                 "io.mantisrx:mantis-control-plane-core"
             ],
-            "locked": "2.11.2"
+            "locked": "2.12.0"
         },
         "junit:junit": {
             "locked": "4.11"
@@ -1172,7 +1193,8 @@
         },
         "org.apache.hadoop:hadoop-common": {
             "firstLevelTransitive": [
-                "io.mantisrx:mantis-control-plane-client"
+                "io.mantisrx:mantis-control-plane-client",
+                "io.mantisrx:mantis-runtime-loader"
             ],
             "locked": "2.7.7"
         },
