+++ mantis-server/mantis-server-worker/src/main/resources/META-INF/services/io.mantisrx.runtime.loader.ITask
